+++ mocha.start.js
@@ -10,8 +10,8 @@ var JSData = require('./');
 
 var store, DSUtils, DSErrors;
 
-function objectsEqual (a, b, msg) {
-  assert.equal(JSON.stringify(a), JSON.stringify(b), msg || 'Expected objects or arrays to be equal');
+assert.objectsEqual = function (a, b, msg) {
+  assert.deepEqual(JSON.parse(JSON.stringify(a)), JSON.parse(JSON.stringify(b)), msg || 'Expected objects or arrays to be equal');
 }
 var lifecycle = {};
 
@@ -42,8 +42,7 @@ var globals = module.exports = {
   TYPES_EXCEPT_FUNCTION: ['string', 123, 123.123, null, undefined, {}, [], true, false],
   assert: assert,
   sinon: sinon,
-  store: undefined,
-  objectsEqual: objectsEqual
+  store: undefined
 };
 
 var test = new mocha();
@@ -150,6 +149,10 @@ beforeEach(function () {
         comment: {
           localField: 'comments',
           foreignKey: 'approvedBy'
+        },
+        group: {
+          localField: 'groups',
+          foreignKeys: 'userIds'
         }
       },
       hasOne: {
@@ -163,6 +166,18 @@ beforeEach(function () {
           parent: true,
           localKey: 'organizationId',
           localField: 'organization'
+        }
+      }
+    }
+  });
+
+  globals.Group = global.Group = store.defineResource({
+    name: 'group',
+    relations: {
+      hasMany: {
+        user: {
+          localField: 'users',
+          localKeys: 'userIds'
         }
       }
     }
@@ -302,6 +317,16 @@ beforeEach(function () {
     id: 18,
     organizationId: 15,
     name: 'test user 18'
+  };
+  globals.group1 = global.group1 = {
+    name: 'group 1',
+    id: 1,
+    userIds: [10]
+  };
+  globals.group2 = global.group2 = {
+    name: 'group 2',
+    id: 2,
+    userIds: [10]
   };
   globals.organization15 = global.organization15 = {
     name: 'Another Test Corp',